build version from git

+++ build.sh
@@ -1,13 +1,20 @@
 #!/bin/bash
 
+PATH=$PATH:/c/Python27/
 VER=`git describe`
+ARCH=`uname -m`
 
-../pyinstaller-1.5.1/pyinstaller.py --onefile context
+echo "VERSION='$VER'" > ctx_ver.py
+echo "!define VERSION '${VER}'" > ctx_ver.nsh
 
-rm -rf context-$VER
-mkdir context-$VER
-cp -rv api context-$VER/
-cp -rv images context-$VER/
-cp -rv docs context-$VER/
-cp dist/* context-$VER/
-tar cvzf context-$VER-`uname -i`.tgz --exclude "*.pyc" context-$VER
+../pyinstaller-1.5.1/pyinstaller.py --onefile --windowed --upx --icon images/boomtools.ico context
+
+if [ "`uname -s`" = "Linux" ] ; then
+	rm -rf context-$VER
+	mkdir context-$VER
+	cp -rv api context-$VER/
+	cp -rv images context-$VER/
+	cp -rv docs context-$VER/
+	cp dist/* context-$VER/
+	tar cvzf context-$VER-$ARCH.tgz --exclude "*.pyc" context-$VER
+fi
